pyproject.toml
[build-system]
requires = ["poetry>=1.1.13"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 99
target-version = ['py37']
skip-string-normalization = true

[tool.coverage.run]
branch = true
source = ["pgpubsub"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise AssertionError",
    "raise NotImplementedError",
    "pass",
    "pytest.mark.skip"
]
show_missing = true
fail_under = 100

[tool.poetry]
name = "django-pgpubsub"
packages = [
  { include = "pgpubsub" }
]
version = "0.0.0"
description = "Djagno integration with Postgres LISTEN and NOTIFY"
authors = ["Opus 10 Engineering"]
classifiers = [
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3 :: Only",
  "Framework :: Django",
]
license = "BSD-3-Clause"
readme = "README.rst"
homepage = "https://github.com/Opus10/django-pgpubsub"
repository = "https://github.com/Opus10/django-pgpubsub"
documentation = "https://django-pgpubsub.readthedocs.io"

[tool.poetry.dependencies]
python = ">=3.7.0,<4"
django = ">=2"
django-pgtrigger = ">=2.4.0"

[tool.poetry.dev-dependencies]
black = "21.5b1"
dj-database-url = "0.5.0"
flake8 = "3.9.2"
flake8-bugbear = "22.1.11"
flake8-comprehensions = "3.8.0"
flake8-import-order = "0.18.1"
flake8-logging-format = "0.6.0"
flake8-mutable = "1.2.0"
git-tidy = "1.1.2"
packaging = ">=19.2"
pip = "*"
pre-commit = "2.13.0"
psycopg2-binary = "2.9.3"
pytest = "7.0.0"
pytest-cov = "3.0.0"
pytest-dotenv = "0.5.2"
pytest-django = "4.5.2"
pytest-mock = "3.7.0"
requests = "2.25.1"
Sphinx = "4.4.0"
sphinx-rtd-theme = "1.0.0"
temple = "*"
tox = "3.24.5"
zipp = "3.4.1"

[tool.pytest.ini_options]
xfail_strict = true
addopts = "--reuse-db"
testpaths = "pgpubsub/tests"
norecursedirs = ".venv"
DJANGO_SETTINGS_MODULE = "settings"
